in production mode user cannot delete such things as menu, menu_item, page & content block

+++ protected/modules/menu/controllers/MenuitemBackendController.php
@@ -184,6 +184,13 @@ class MenuitemBackendController extends yupe\components\controllers\BackControll
      */
     public function actionDelete($id)
     {
+        if (!YII_DEBUG) {
+            throw new CHttpException(
+                403,
+                Yii::t('MenuModule.menu', 'Cannot delete in Production Mode!')
+            );
+        }
+
         if (Yii::app()->getRequest()->getIsPostRequest()) {
             // we only allow deletion via POST request
             $this->loadModel($id)->deleteWithChild();
